backend/requirements/base.txt
# Django Core
Django>=4.2,<5.0
django-environ>=0.11.2

# GraphQL
graphene-django>=3.1.5
django-graphql-jwt>=0.4.0
graphene>=3.3

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# Authentication & Security
PyJWT>=2.8.0
argon2-cffi>=23.1.0
django-cors-headers>=4.3.1

# Caching & Background Tasks
django-redis>=5.4.0
celery[redis]>=5.3.6
redis>=5.0.1

# AI & ML Dependencies
sentence-transformers>=2.2.2
scikit-learn>=1.3.2
numpy>=1.26.2
pandas>=2.1.4

# OCR & Face Recognition
pytesseract>=0.3.10
Pillow>=10.1.0
face-recognition>=1.3.0
opencv-python-headless>=4.8.1.78

# NLP
nltk>=3.8.1

# File Handling
python-magic>=0.4.27

# Utilities
django-filter>=23.5
django-extensions>=3.2.3
uuid>=1.30
requests>=2.31.0

# ASGI
uvicorn>=0.25.0
